BiG-SCAPE will now create SVG figures for each input BGC Necessary files: ArrowerSVG.py Optional files: domains_color_file.tsv, gene_color_file.tsv and pfam_domain_categories.tsv
All the SVG figures will be put in the <outputfolder>/SVG folder.

The Arrower code will open the input gbk file and the pfd (processed domtable)
files to extract information about the BGC (gene and domain position, locus
length). Colors for genes and domains will be read from the first two optional
files; if not present, they will be created and new colors will be generated.

The pfam_domain_categories.tsv file currently classifies domains into four
categories related to their functionality. If this file is present, a shadow
may be drawn around a gene according to the following logic:
- Core Biosynthetic. Red shadow. It contains at least one Core Bio. domain
- Other Biosynthetic. Orange shadow. It contains at least one Other Bio. domain
and none of the other
- Transporter. Blue shadow. Contains at least one Transporter domain and none
of the other
- Regulator. Green shadow. Contains at least one Regulator domain and none of
the other
This logic may change in the future.


Former-commit-id: 94365d287de989f0fd13014ffa8234f58f4c5e1a